mySTEP.make: new feature to manage Debian-Release specific dependencies
Works as follows:
provide DEBIAN_DEPENDS with dependency information as usual (comma separated list)
add prefix: to define a debian release and add this dependency only to the specific
package

Example:
DEBIAN_DEPENDS="general, wheezy:wheezy-only, jessie:jessie-only"

This generates two packages and places them separated in the staging area (by.
wheezy/jessie in the directory path).

If no colon-specifier is found, the package is placed in the "staging" directory.

Signed-off-by: H. Nikolaus Schaller <hns@goldelico.com>