NSNotificationQueue: process queue in correct order
The naming of prev/next and head/tail is opposite to the typical convention,
hence we must start with the head, but iterate by item->prev to process
the full list.

The current code handles only one notification per call.
This makes notifications getting stuck in the queue and never delivered.

Reported-By: "Felipe A Rodriguez" <far@illumenos.com>
Signed-off-by: H. Nikolaus Schaller <hns@goldelico.com>