mySTEP.make: attempt to build multi-arch packages for supporting also multiple OS releases
The problem to be solved is that we need different toolchains and libary versions
to link against. E.g. libtiff or libjpeg versions (and headers) differ between Jessie, Stretch, Buster, Bullseye.
The same may even be for libobjc.
So we must compile multiple times with different toolchains and keep the .o and .so files separate.
With this we can already build a single framework and debuan package that supports multiple releases
and architectures.
Two issues are not yet well solved:
a) how linking other frameworks or apps does find the release specific .so files
b) how runtime finds the right .so file
especially since we do not want to build every potential variant.
Ideally only those frameworks have multiple variants which really need it.

Signed-off-by: H. Nikolaus Schaller <hns@goldelico.com>