Foundation: NSArray: do not do a deep copyWithZone!
This is old behaviour up to 10.4. See:
https://developer.apple.com/library/archive/releasenotes/Foundation/RN-FoundationOlderNotes/#X10_5Notes
(search for Immutable collections and copy behavior)

A typical code snippet where it makes problems:
NSMutableArray *someArray;
NSEnumerator *e=[[someArray copy] objectEnumerator];
... loop and modify someArray

Here using a default deep copy of the array requires all array members
to implement the copying protocol which is likely not intended and not the case.
An example where it makes such trouble is VectorBoolean

            NSArray *crossings = [[ourEdge.crossings copy] autorelease];

This raises an exception on mySTEP while it works on Cocoa

Signed-off-by: H. Nikolaus Schaller <hns@goldelico.com>