mySTEP.make: retire codename: syntax to define target(s)
There was a special syntax for DEBIAN_DEPENDS etc. like
jessie:package1, stretch:package2, trixie:package3

This had a dual purpose:
a) tell mySTEP.make that we want to build jessie, stretch and trixie variants
b) have variants (codename) dependent dependencies

The purpose a) can me more simply achieved by (optionally) adding
DEBIAN_RELEASES="jessie stretch trixie" to the .qcodeprojec
A fallback is to build for staging which will be copied into
all suites/codenames.

Purpose b) can be more easily achieved by using DEBIAN_RECOMMENDS for
packages that do not exist in every codename. apt-get install will
recognise them but ignore on a specific codename if the package does
not exist (there).

An alterantive is to use e.g. php5 | php7 | php82 to install either
depending on what is available. A small difference to DEBIAN_RECOMMENDS
is that this requires that at least one of the given alternatives really
exists or the full package can not be installed

Signed-off-by: H. Nikolaus Schaller <hns@goldelico.com>